refactor(vote chart): moved vote response chart to separate component
